move normalize_vecs to tensor_utils
