+++ mmedit/models/utils/__init__.py
@@ -6,12 +6,12 @@ from .model_utils import (default_init_weights, generation_init_weights,
                           get_module_device, get_valid_noise_size,
                           get_valid_num_batches, make_layer, set_requires_grad)
 from .sampling_utils import label_sample_fn, noise_sample_fn
-from .tensor_utils import get_unknown_tensor
+from .tensor_utils import get_unknown_tensor, normalize_vecs
 
 __all__ = [
     'default_init_weights', 'make_layer', 'flow_warp',
     'generation_init_weights', 'set_requires_grad', 'extract_bbox_patch',
     'extract_around_bbox', 'get_unknown_tensor', 'noise_sample_fn',
     'label_sample_fn', 'get_valid_num_batches', 'get_valid_noise_size',
-    'get_module_device'
+    'get_module_device', 'normalize_vecs'
 ]
