+++ tests/test_models/test_editors/test_eg3d/test_camera.py
@@ -24,7 +24,6 @@ class TestBaseCamera(TestCase):
     def test_init(self):
         cfg_ = deepcopy(self.default_cfg)
         camera = BaseCamera(**cfg_)
-        # self.assertEqual(camera.device, 'cpu')
         self.assertEqual(camera.sampling_statregy.upper(), 'UNIFORM')
 
         cfg_ = deepcopy(self.default_cfg)
@@ -65,25 +64,6 @@ class TestBaseCamera(TestCase):
         repr_string = repr(camera)
         self.assertNotIn('focal', repr_string)
         self.assertNotIn('fov', repr_string)
-
-    # def test_device(self):
-    #     cfg_ = deepcopy(self.default_cfg)
-    #     camera = BaseCamera(**cfg_)
-    #     return_obj = camera.to('cuda')
-    #     self.assertEqual(camera.device, 'cuda')
-    #     self.assertIsInstance(return_obj, BaseCamera)
-
-    #     return_obj = camera.cpu()
-    #     self.assertEqual(camera.device, 'cpu')
-    #     self.assertIsInstance(return_obj, BaseCamera)
-
-    #     return_obj = camera.cuda()
-    #     self.assertEqual(camera.device, 'cuda')
-    #     self.assertIsInstance(return_obj, BaseCamera)
-
-    #     return_obj = camera.cuda(3)
-    #     self.assertEqual(camera.device, 'cuda:3')
-    #     self.assertIsInstance(return_obj, BaseCamera)
 
     def test_sample_intrinsic(self):
         target = [1.025390625, 0.0, 0.5, 0.0, 1.025390625, 0.5, 0.0, 0.0, 1.0]
@@ -153,7 +133,6 @@ class TestUniformCamera(TestCase):
     def test_init(self):
         cfg_ = deepcopy(self.default_cfg)
         camera = UniformCamera(**cfg_)
-        # self.assertEqual(camera.device, 'cpu')
         self.assertEqual(camera.sampling_statregy.upper(), 'UNIFORM')
 
     def test_repr(self):
@@ -177,7 +156,6 @@ class TestGaussianCamera(TestCase):
     def test_init(self):
         cfg_ = deepcopy(self.default_cfg)
         camera = GaussianCamera(**cfg_)
-        # self.assertEqual(camera.device, 'cpu')
         self.assertEqual(camera.sampling_statregy.upper(), 'GAUSSIAN')
 
     def test_repr(self):
